Type casting has tests
