+++ type.ts
@@ -1,0 +1,11 @@
+export const UNKNOWN = 'unknown';
+
+export enum Type {
+  Integer = 'integer',
+  Real = 'real',
+  Boolean = 'boolean',
+  String = 'string',
+  Exception = 'Exception',
+  Nil = 'nil',
+  Any = 'any',
+}
